# doc/api.yaml
openapi: 3.0.0
info:
  title: E-commerce API
  description: This API handles user authentication, product management, cart operations, and order processing for an e-commerce platform.
  version: 1.0.0
servers:
  - url: 'http://localhost:3000'
    description: Development server
paths:
...
  /products:
    get:
      summary: List all products
      tags:
        - Products
      responses:
        '200':
          description: List of all products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
    post:
      summary: Create a new product
      tags:
        - Products
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '500':
          description: Error creating product
...
  /cart/{userId}:
    get:
      summary: Retrieve a user's cart
      tags:
        - Carts
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: Unique identifier of the user
      responses:
        '200':
          description: Cart details retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CartItem'
        '404':
          description: Cart not found
        '500':
          description: Server error
  /cart/{userId}/items:
    post:
      summary: Add an item to a user's cart
      tags:
        - Carts
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
          description: Unique identifier of the user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                productId:
                  type: integer
                quantity:
                  type: integer
      responses:
        '201':
          description: Item added to cart
        '500':
          description: Error adding item to cart
...
  /token:
    post:
      summary: Refresh access token using a refresh token
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
      responses:
        '200':
          description: Access token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
        '401':
          description: No refresh token provided
        '403':
          description: Invalid or expired refresh token
        '500':
          description: Server error
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        stock:
          type: integer
    CartItem:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
...